Move NixError from exception package to utility.  Clean up minor PEP 8 defects.
