+++ src/utility/print_table.py
@@ -14,6 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+
 class PrintTable(object):
     def __init__(self, *args):
         self._data = [list(args)]
