+++ test/config/script/test_updateScript.py
@@ -14,6 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+
 class TestUpdateScript:
     def test_update(self, config_valid):
         pass
